pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "crypto-bot"
version = "1.0.0"
description = "Cryptocurrency trading system with backtesting, live trading, and ML signals"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Trading System Team"},
]
requires-python = ">=3.8"
keywords = ["crypto", "trading", "bitcoin", "ethereum", "backtesting", "machine-learning"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "flask>=2.0.0",
    "flask-socketio>=5.0.0",
    "python-socketio>=5.0.0",
    "python-engineio>=4.0.0",
    "requests>=2.25.0",
    "pandas>=1.2.0",
    "numpy>=1.19.0",
    "scipy>=1.5.0",
    "scikit-learn>=0.24.0",
    "matplotlib>=3.3.0",
    "seaborn>=0.11.0",
    "python-dotenv>=0.19.0",
    "pyyaml>=5.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.12.0",
    "black>=21.0",
    "flake8>=3.9.0",
    "mypy>=0.900",
]
test = [
    "pytest>=6.0.0",
    "pytest-cov>=2.12.0",
]

[project.urls]
Homepage = "https://github.com/your-username/crypto-bot"
Repository = "https://github.com/your-username/crypto-bot.git"
"Bug Tracker" = "https://github.com/your-username/crypto-bot/issues"

[tool.setuptools]
packages = ["crypto_bot"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
crypto_bot = ["config/*.json", "models/*.pkl"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.coverage.run]
source = ["src/crypto_bot"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
